requirements.txt
pydantic>=1.7.3
gitpython>=3.1.12
requests>=2.25.1
joblib>=1.0.1